Azure: Developer deployment
- Redo the azurdeploy.json to be like the dev template
without the 1 PubIp and 1 data disk on head node optimization
we use for customr deployments. For eng/dev/test we
would like to be able to access all instances easily, until
we have bastion host infra checked in. This also means we
support single node cluster again for azuredpeloy.json

- By default use our own Xcalar CA for https instead of
getting rate limited on LetsEncrypt

- On AMP use pregenerated, legitimate LE SAN certs for the
first 100 customers, then revert to asking LE for certs. Since
we can't pass a signed url at AMP deploy time, the file needs
to be publically avaible. Use gpg to encrypt the file and
pass the password in 'protected' settings of template.

- Use xfs non-cache device for NFS share as this reduces bootstrap
time by 9-10min. We still use ext4 on the local cache ssd
because the swapfile can't quickly be created on xfs. You
need to use dd instead of being able to use fallocate

- Install azure-cli package and create default storage
container 'customer'

+++ azure/bootstrap/ampBootstrap.sh
@@ -2,39 +2,42 @@
 
 echo "Starting bootstrap at `date`"
 
-echo "$@" | tee args.txt
-
 INSTALLER_SERVER="https://zqdkg79rbi.execute-api.us-west-2.amazonaws.com/stable/installer"
 LICENSE_SERVER="https://zd.xcalar.net/license/api/v1.0/marketplacedeploy"
 HTML="http://pub.xcalar.net/azure/dev/html-4.tar.gz"
 XCALAR_ADVENTURE_DATASET="http://pub.xcalar.net/datasets/xcalarAdventure.tar.gz"
-ZONE="Z38PTJSFJD11PH"
-SUBDOMAIN="azure.xcalar.cloud"
+SUBDOMAIN="azure.xcalar.io"
+CONTAINER="customer"
+export AWS_HOSTED_ZONE_ID="Z1B5U8RYSOLN6J" # for lego
 export AWS_DEFAULT_REGION=us-west-2
 
 # By default use the LetsEncrypt staging server so we don't trigger
 # CA limits for this domain
-CASERVER="https://acme-staging.api.letsencrypt.org/directory"
+CASTAGING="https://acme-staging.api.letsencrypt.org/directory"
+CASERVER="$CASTAGING"
 
-while getopts "a:b:c:d:e:f:i:n:l:u:r:s:v:w:x:y:z:" optarg; do
+while getopts "a:b:c:d:e:f:g:i:n:l:u:r:p:s:t:v:w:x:y:z:" optarg; do
     case "$optarg" in
         a) SUBDOMAIN="$OPTARG";;
-        b) ZONE="$OPTARG";;
+        b) export AWS_HOSTED_ZONE_ID="$OPTARG";;
         c) CLUSTER="$OPTARG";;
         d) DNSLABELPREFIX="$OPTARG";;
         e) export AWS_ACCESS_KEY_ID="$OPTARG";;
         f) export AWS_SECRET_ACCESS_KEY="$OPTARG";;
+        g) PASSWORD="$OPTARG";;
         i) INDEX="$OPTARG";;
         n) COUNT="$OPTARG";;
         l) LICENSE="$OPTARG";;
         u) INSTALLER_URL="$OPTARG";;
         r) CASERVER="$OPTARG";;
+        p) PEM_URL="$OPTARG";;
+        t) CONTAINER="$OPTARG";;
         s) NFSMOUNT="$OPTARG";;
         v) ADMIN_EMAIL="$OPTARG";;
         w) ADMIN_USERNAME="$OPTARG";;
         x) ADMIN_PASSWORD="$OPTARG";;
-        y) STORAGE_ACCOUNT_NAME="$OPTARG";;
-        z) STORAGE_ACCESS_KEY="$OPTARG";;
+        y) export AZURE_STORAGE_ACCOUNT="$OPTARG";;
+        z) export AZURE_STORAGE_ACCESS_KEY="$OPTARG"; export AZURE_STORAGE_KEY="$OPTARG";;
         --) break;;
         *) echo >&2 "Unknown option $optarg $OPTARG";; # exit 2;;
     esac
@@ -63,13 +66,13 @@ clean_fstab () {
 # device is already mounted somewhere else, it is unmounted. *CAREFUL* when calling
 # this function, it will destroy the specified device.
 mount_device () {
-    test $# -eq 2 || return 1
+    test $# -ge 2 || return 1
     test -n "$1" && test -n "$2" || return 1
-    local PART= DEV="${2%[1-9]}" retry=
-    if PART="$(set -o pipefail; findmnt -n $1 | awk '{print $2}')"; then
-        local OLDMOUNT="$(findmnt -n $1 | awk '{print $1}')"
-        if [ "$PART" != "$2" ] || [ -z "$OLDMOUNT" ]; then
-            echo >&2 "Bad mount $1 on device $2. Bailing." >&2
+    local PART= MOUNT="$1" PARTIN="$2" DEV="${2%[1-9]}" LABEL="$3" FSTYPE="${4:-ext4}"
+    if PART="$(set -o pipefail; findmnt -n $MOUNT | awk '{print $2}')"; then
+        local OLDMOUNT="$(findmnt -n $MOUNT | awk '{print $1}')"
+        if [ "$PART" != "$PARTIN" ] || [ -z "$OLDMOUNT" ]; then
+            echo >&2 "Bad mount $MOUNT on device $PARTIN. Bailing." >&2
             return 1
         fi
         umount $OLDMOUNT
@@ -77,32 +80,69 @@ mount_device () {
     # If there's already a partition table, you need to sgdisk it twice
     # because it 'fails' the first time. sgdisk aligns the partition for you
     # -n1 creates an aligned partition using the entire disk, -t1 sets the
-    # partition type to 'Linux filesystem' and -c1 sets the label to 'SSD'
-    sgdisk -Zg -n1:0:0 -t1:8300 -c1:SSD $DEV || sgdisk -Zg -n1:0:0 -t1:8300 -c1:SSD $DEV
+    # partition type to 'Linux filesystem' and -c1 sets the label to 'LABEL'
+    sgdisk -Zg -n1:0:0 -t1:8300 -c1:$LABEL $DEV || sgdisk -Zg -n1:0:0 -t1:8300 -c1:$LABEL $DEV
     test $? -eq 0 || return 1
     sync
+    local retry=
     for retry in $(seq 5); do
         sleep 5
-        # Must use -F[orce] because the partition may have already existed with a valid
-        # file system. sgdisk doesn't earase the partitioning information, unlike parted/fdisk.
-        # lazy_itable_init=0,lazy_journal_init=0 take too long on Azure
-        time mkfs.ext4 -F -m 0 -E discard $2 && break
+        if [ "$FSTYPE" = xfs ]; then
+            time mkfs.xfs -f $PARTIN && break
+        elif [ "$FSTYPE" = ext4 ]; then
+            # Must use -F[orce] because the partition may have already existed with a valid
+            # file system. sgdisk doesn't earase the partitioning information, unlike parted/fdisk.
+            # lazy_itable_init=0,lazy_journal_init=0 take too long on Azure
+            time mkfs.ext4 -F -m 0 -E nodiscard $PARTIN && break
+        fi
     done
     test $? -eq 0 || return 1
-    clean_fstab $2 && \
-    mkdir -p $1 && \
-    echo "$2   $1      ext4        defaults,discard,relatime,nobarrier  0   0" | tee -a /etc/fstab
-    mount $1
+    local UUID="$(blkid -s UUID $PARTIN -o value)"
+    clean_fstab $UUID && \
+    clean_fstab $PARTIN && \
+    mkdir -p $MOUNT && \
+    if [ "$FSTYPE" = xfs ]; then
+        echo "UUID=$UUID   $MOUNT      xfs         defaults,discard,relatime,nobarrier  0   0" | tee -a /etc/fstab
+    elif [ "$FSTYPE" = ext4 ]; then
+        echo "UUID=$UUID   $MOUNT      ext4        defaults,discard,relatime,nobarrier  0   0" | tee -a /etc/fstab
+    fi
+    mount $MOUNT
+}
+
+download_cert () {
+    local gpgfile="$(basename "$1")" try=0
+    local tarfile="$(basename "$gpgfile" .gpg)"
+    local pemfile= keyfile=
+    aws s3 cp "$1" "$gpgfile" >&2 && \
+    while [ $try -lt 10 ]; do
+        if [[ $gpgfile =~ \.gpg$ ]]; then
+            echo "$2"| gpg --passphrase-fd 0 --batch --quiet --yes --no-tty -d "$gpgfile" > "$tarfile"
+        fi && \
+        pemfile="$(tar tf "$tarfile" | grep -E '\.(pem|crt)' | head -1)" && \
+        keyfile="$(tar tf "$tarfile" | grep -E '\.key' | head -1)" && \
+        tar zxf "$tarfile" && \
+        echo "${PWD}/${pemfile}" && \
+        return 0
+        try=$((try+1))
+        sleep 5
+    done
+    return 1
 }
 
 lego_register_domain() {
-    safe_curl -L https://github.com/xenolf/lego/releases/download/v0.4.0/lego_linux_amd64.tar.xz | \
-        tar Jxvf - --no-same-owner lego_linux_amd64
-    mv lego_linux_amd64 /usr/local/bin/lego
-    chmod +x /usr/local/bin/lego
-    setcap cap_net_bind_service=+ep /usr/local/bin/lego
-    lego -d "$1" --server "${CASERVER}" --dns route53 --accept-tos --email "${ADMIN_EMAIL}" run
-    if [ $? -ne 0 ]; then
+    if ! test -e /usr/local/bin/lego; then
+        safe_curl -L https://github.com/xenolf/lego/releases/download/v0.4.0/lego_linux_amd64.tar.xz | \
+            tar Jxvf - --no-same-owner lego_linux_amd64
+        mv lego_linux_amd64 /usr/local/bin/lego
+        chmod +x /usr/local/bin/lego
+        setcap cap_net_bind_service=+ep /usr/local/bin/lego
+    fi
+    local caserver="$1" domains=() domain=
+    shift
+    for domain in "$@"; do
+        domains+=(-d $domain)
+    done
+    if ! lego --server "$caserver" "${domains[@]}" --dns route53 --accept-tos --email "${ADMIN_EMAIL}" run; then
         echo >&2 "Failed to acquire certificate"
         return 1
     fi
@@ -113,19 +153,51 @@ lego_register_domain() {
 
 # Create a resource record that points xd-standard-amit-0.westus2.cloud.azure.com -> yourprefix.azure.xcalar.cloud
 aws_route53_record () {
-    local CNAME="$1" NAME="$2" rrtmp="$(mktemp /tmp/rrsetXXXXXX.json)"
+    local CNAME="$1" NAME="$2" rrtmp="$(mktemp /tmp/rrsetXXXXXX.json)" change_id=
     cat > $rrtmp <<EOF
-    { "HostedZoneId": "$ZONE", "ChangeBatch": { "Comment": "Adding $CNAME",
+    { "HostedZoneId": "$AWS_HOSTED_ZONE_ID", "ChangeBatch": { "Comment": "Adding $CNAME",
       "Changes": [ {
         "Action": "UPSERT",
-          "ResourceRecordSet": { "Name": "$NAME", "Type": "CNAME", "TTL": 300,
+          "ResourceRecordSet": { "Name": "$NAME", "Type": "CNAME", "TTL": 60,
             "ResourceRecords": [ { "Value": "$CNAME" } ] } } ] } }
 EOF
-    aws route53 change-resource-record-sets --cli-input-json file://${rrtmp}
+    change_id="$(set -o pipefail; aws route53 change-resource-record-sets --cli-input-json file://${rrtmp} | jq -r '.ChangeInfo.Id' | cut -d'/' -f3)"
+    if [ $? -eq 0 ] && [ "$change_id" != "" ]; then
+        echo "$change_id"
+        return 0
+    fi
+    return 1
+}
+
+pem_san_list () {
+    openssl x509 -noout -text -in "$1" | grep 'DNS:' | sed -r 's/ DNS://g; s/^[\t ]+//g; s/,/\n/g'
+}
+
+aws_route53_list () {
+    aws route53 list-resource-record-sets --hosted-zone-id "$AWS_HOSTED_ZONE_ID" --query 'ResourceRecordSets[].Name' --output text | tr '\t' '\n' | grep '\.'${SUBDOMAIN} | sort
+}
+
+dns_find_slot () {
+    aws_route53_list > zone_records.txt
+    pem_san_list "$1" > pem_san.txt
+    if grep -q '\*\.'${SUBDOMAIN} pem_san.txt; then
+        echo "${DNSLABELPREFIX}.${SUBDOMAIN}"
+        return 0
+    fi
+    comm -13 zone_records.txt pem_san.txt > allowed.txt
+    local precert="$(comm -13 zone_records.txt pem_san.txt | head -1 | tee -a dnsname.txt)"
+    if [ -z "$precert" ]; then
+        return 1
+    fi
+    echo "$precert"
 }
 
 setenforce Permissive
 sed -i -e 's/^SELINUX=enforcing.*$/SELINUX=permissive/g' /etc/selinux/config
+
+# AzureCLI (ref: https://docs.microsoft.com/en-us/cli/azure/install-azure-cli?view=azure-cli-latest)
+rpm --import https://packages.microsoft.com/keys/microsoft.asc
+echo -e "[azure-cli]\nname=Azure CLI\nbaseurl=https://packages.microsoft.com/yumrepos/azure-cli\nenabled=1\ngpgcheck=1\ngpgkey=https://packages.microsoft.com/keys/microsoft.asc" > /etc/yum.repos.d/azure-cli.repo
 
 yum makecache fast
 
@@ -134,7 +206,7 @@ rpm -e hypervkvpd || true
 yum install -y microsoft-hyper-v
 
 yum install -y nfs-utils epel-release parted gdisk curl
-yum install -y jq python-pip awscli
+yum install -y jq python-pip awscli azure-cli
 
 # For CIFS
 yum install -y samba-client samba-common cifs-utils
@@ -212,14 +284,14 @@ LOCATION="$(<metadata.json jq -r '.compute.location')"
 # On some Azure instances /mnt/resource comes premounted but not aligned properly
 RESOURCEDEV="$(findmnt -n /mnt/resource | awk '{print $2}')"
 if [ -n "$RESOURCEDEV" ]; then
-    mount_device /mnt/resource $RESOURCEDEV
+    mount_device /mnt/resource $RESOURCEDEV SSD ext4
     INSTANCESTORE=/mnt/resource
 fi
 
 # Format and mount additional SSD, and prefer to use that
 for DEV in /dev/sdb /dev/sdc /dev/sdd; do
     if test -b ${DEV} && ! test -b "${DEV}1"; then
-        mount_device /mnt/ssd  "${DEV}1"
+        mount_device /mnt/ssd  "${DEV}1" SSD2 xfs
         LOCALSTORE=/mnt/ssd
         break
     fi
@@ -273,6 +345,8 @@ if [ "$HOSTNAME" = "$NFSHOST" ]; then
     fi
 fi
 
+
+### Install Xcalar
 if [ -f "installer.sh" ]; then
     if ! bash -x installer.sh --nostart --caddy; then
         echo >&2 "ERROR: Failed to run installer"
@@ -285,13 +359,67 @@ if [ -f "installer.sh" ]; then
     setcap cap_net_bind_service=+ep $XLRDIR/bin/caddy
 fi
 
-if [ -r /etc/default/xcalar ]; then
-    echo "" >> /etc/default/xcalar
-    echo "## Azure Blob Storage config" >> /etc/default/xcalar
-    echo "AZURE_STORAGE_ACCOUNT=$STORAGE_ACCOUNT_NAME" >> /etc/default/xcalar
-    echo "AZURE_STORAGE_ACCESS_KEY=$STORAGE_ACCESS_KEY" >> /etc/default/xcalar
-    echo "export AZURE_STORAGE_ACCOUNT AZURE_STORAGE_ACCESS_KEY" >> /etc/default/xcalar
-    . /etc/default/xcalar
+# Download PEM
+if [ -n "$PEM_URL" ] && [ -n "$PASSWORD" ]; then
+    if PEM="$(download_cert "$PEM_URL" "$PASSWORD")"; then
+        cp "$PEM" /etc/xcalar/cert.pem && \
+        cp "$(dirname $PEM)/$(basename $PEM .pem).key" /etc/xcalar/cert.key && \
+        chmod 0444 /etc/xcalar/cert.pem /etc/xcalar/cert.key && \
+        PEM="/etc/xcalar/cert.pem" || {
+            PEM=""
+            echo >&2 "Claimed to have downloaded cert, but there isn't one"
+        }
+    fi
+fi
+
+# Returns relative date in az compatible format
+# ex: format_expiry "10 days" -> 2017-10-01T1200Z
+az_format_expiry () {
+    date -u -d "$1" +'%Y-%m-%dT%H:%MZ'
+}
+
+# generate an az storage account sas token given an expiry date relative from now (eg, "10 days")
+# more info: az storage account generate-sas --help
+# quick ref: services -> (b)lob, (f)ile ..
+#            resource-types -> (s)ervice, (c)ontainer, (o)bject
+#            permissions -> (a)dd, (c)create, ..
+az_storage_sas () {
+    az storage account generate-sas --services bfqt --resource-types sco --permissions acdlpruw --expiry $(az_format_expiry "$1") --output tsv
+}
+
+# TODO: Should store this instead of AZURE_*_KEY
+export AZURE_STORAGE_SAS_TOKEN="$(az_storage_sas '90 days')"
+if touch /etc/azure; then
+    # Only allow access by root
+    chmod 0600 /etc/azure
+    echo "## Azure Blob Storage config" >> /etc/azure
+    echo "AZURE_STORAGE_ACCOUNT=$AZURE_STORAGE_ACCOUNT" >> /etc/azure
+    echo "AZURE_STORAGE_ACCESS_KEY=$AZURE_STORAGE_ACCESS_KEY" >> /etc/azure
+    echo "AZURE_STORAGE_KEY=$AZURE_STORAGE_KEY" >> /etc/azure
+    echo "AZURE_STORAGE_SAS_TOKEN=\"$AZURE_STORAGE_SAS_TOKEN\"" >> /etc/azure
+    echo "export AZURE_STORAGE_ACCOUNT AZURE_STORAGE_ACCESS_KEY AZURE_STORAGE_KEY AZURE_STORAGE_SAS_TOKEN" >> /etc/azure
+    if [ -r /etc/default/xcalar ]; then
+        # should filter out _KEY and only keep ACCOUNT_NAME and SAS_TOKEN in
+        # /etc/default/xcalar. Since this file contains secrets, remove world
+        # readable bit
+        chmod 0640 /etc/default/xcalar
+        cat /etc/azure | tee -a /etc/default/xcalar >/dev/null
+        . /etc/default/xcalar
+    else
+        . /etc/azure
+    fi
+fi
+
+# Only have head node create the container
+if [ "$INDEX" = 0 ] && [ -n "$CONTAINER" ]; then
+    # Don't strictly need to pass the account name and sas token as they're in env vars, just here
+    # for reference should we decide to remove the global env vars
+    CONTAINER_CREATED=$(az storage container create --account-name "$AZURE_STORAGE_ACCOUNT" --sas-token "$AZURE_STORAGE_SAS_TOKEN" --name $CONTAINER --query 'created')
+    if [ "$CONTAINER_CREATED" = true ]; then
+        echo "Created container $CONTAINER"
+    else
+        echo "Failed to create container $CONTAINER"
+    fi
 fi
 
 # Generate a list of all cluster members
@@ -303,9 +431,20 @@ done
 
 # Register domain
 CNAME="${DNSLABELPREFIX}-${INDEX}.${LOCATION}.cloudapp.azure.com"
-XCE_DNS="${DNSLABELPREFIX}.${SUBDOMAIN}"
+
 DEPLOYED_URL=""
-if [ "$INDEX" = 0 ]; then
+XCE_DNS=""
+if [ "$PUBLICIPV4" != "" ]; then
+    if [ "$INDEX" = 0 ]; then
+        if [ -n "$PEM" ]; then
+            XCE_DNS="$(dns_find_slot "$PEM")"
+        else
+            XCE_DNS="${DNSLABELPREFIX}.${SUBDOMAIN}"
+        fi
+    fi
+    if [ -z "$XCE_DNS" ]; then
+        XCE_DNS="${DNSLABELPREFIX}-${INDEX}.${SUBDOMAIN}"
+    fi
     aws_route53_record "${CNAME}" "${XCE_DNS}"
     (
     echo "https://${XCE_DNS}:443 {"
@@ -317,7 +456,14 @@ if [ "$INDEX" = 0 ]; then
     mv /etc/xcalar/Caddyfile.$$ /etc/xcalar/Caddyfile
     # Have to add the -agree flag or caddy asks us interactively
     sed -i -e 's/caddy -quiet/caddy -quiet -agree/g' /etc/xcalar/supervisor.conf
-    if lego_register_domain "${XCE_DNS}"; then
+    if [ -n "$PEM" ]; then
+        sed -i -e "s|tls.*$|tls /etc/xcalar/cert.pem /etc/xcalar/cert.key|g" /etc/xcalar/Caddyfile
+        chmod 0644 /etc/xcalar/cert.*
+        DEPLOYED_URL="https://$XCE_DNS"
+    elif [ "$INDEX" = 0 ] && lego_register_domain "${CASERVER}" "${XCE_DNS}"; then
+        sed -i -e "s|tls.*$|tls /etc/xcalar/${XCE_DNS}.crt /etc/xcalar/${XCE_DNS}.key|g" /etc/xcalar/Caddyfile
+        DEPLOYED_URL="https://$XCE_DNS"
+    elif lego_register_domain "${CASTAGING}" "${XCE_DNS}"; then
         sed -i -e "s|tls.*$|tls /etc/xcalar/${XCE_DNS}.crt /etc/xcalar/${XCE_DNS}.key|g" /etc/xcalar/Caddyfile
         DEPLOYED_URL="https://$XCE_DNS"
     else
@@ -338,7 +484,7 @@ fi
 # Generate /etc/xcalar/default.cfg
 (
 if [ $COUNT -eq 1 ]; then
-    ${XLRDIR}/scripts/genConfig.sh /etc/xcalar/template.cfg - "$HOSTNAME"
+    ${XLRDIR}/scripts/genConfig.sh /etc/xcalar/template.cfg - localhost
 else
     ${XLRDIR}/scripts/genConfig.sh /etc/xcalar/template.cfg - "${MEMBERS[@]}"
 fi
